CA-101699: Run VM.copy on the master
The VDI.copy subtask(s) created by VM.copy will be forwarded to suitable
hosts, so attempting to forward VM.copy is an unnecessary layer of
complexity.

Besides kicking off VDI.copy tasks, VM.copy mostly just creates objects
in the database and watches for events on those VDI.copy tasks, both of
which are most efficient on the master.

Signed-off-by: John Else <john.else@citrix.com>