CP-5597: Fix initial creation of GPU groups for PGPUs
PGPU.set_GPU_group was checking that the source group wasn't left with in a bad
state (i.e. with VGPUs and no PGPUs) but in the case where the group is first
created to put a newly found PGPU in the 'source group' for the newly
discovered PGPU is NULL reference. So we only look for this condition to be met
when a PGPU is being moved between groups and not initially assigned a group.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>