CP-5596: Have VM_start fail if vgpu fails to start
Factored out the code to start Qemu into a function that forks a daemon and
waits for it to finish initialising (with a write to XenStore). This has then
be reused to kick off the vgpu emulator binary which can now fail in the same
way if it doesn't initialise. vgpu is kicked off in conjuction with Qemu since
it's part of the device model.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>