CP-14804: Modify how the datamodel detects removed API messages
Previously, if any of an API message's lifecycle transitions was
Removed, we would treat the message as Removed and autogenerate code to
throw the MESSAGE_REMOVED exception.

Now we check that the last transition is Removed. This allows an API
message to be Removed in one release, and then re-Published in a later
release.

Signed-off-by: John Else <john.else@citrix.com>