Pool.get_license_state: don't make assumptions about editions in xapi
When comparing the editions of the hosts in the pool to compute the "pool
edition", xapi must not assume any knowledge about which editions exist and how
to compare them. It must ask v6d for this information.

We now need a more precise definition of "pool expiry". We take this to be the
earliest expiry date of all hosts that are licensed at the pool edition. That
is, all licenses that are higher than the pool edition are ignored.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>